requirements.txt
requests
futures
flask
python-dotenv
watson-developer-cloud
